feat(serwalker): implemented tag-decoding macros, declared all fixed-size alloc cluster types, implemented parse_clusters, implemented a macro to declare the ClassId enum and TryFrom for u32 for said enum

+++ crates/flutterdec-serwalker/src/raw_object/mod.rs
@@ -1,4 +1,4 @@
-type Smi = i64; // Using i64 for Smi fields to be decompressed
+type Smi = i32; // Using i64 for Smi fields to be decompressed
 
 // --- Fixed-Size Objects with defined fields ---
